// src/shared/types/technologies.types.ts
export enum Technology {
  ACCESSIBILITY = "Accessibility",
  ANGULAR = "Angular",
  AZURE = "Azure",
  BEM = "BEM",
  BOOTSTRAP = "Bootstrap",
  CSS = "CSS",
  CURSOR = "Cursor",
  CYPRESS = "Cypress",
  EXPRESS_JS = "Express.js",
  FIGMA = "Figma",
  GEMINI = "Gemini",
  GIT = "Git",
  GITHUB = "GitHub",
  GITHUB_ACTIONS = "GitHub Actions",
  GITHUB_COPILOT = "GitHub Copilot",
  GITLAB = "GitLab",
  HTML = "HTML",
  ILLUSTRATOR = "Illustrator",
  INDESIGN = "InDesign",
  JASMINE = "Jasmine",
  JAVA = "Java",
  JEST = "Jest",
  JIRA = "Jira",
  JAVASCRIPT = "JavaScript",
  MONGODB = "MongoDB",
  NESTJS = "NestJS",
  NEXT_JS = "Next.js",
  NODE_JS = "Node.js",
  NX = "Nx",
  PERFORMANCE = "Performance",
  PLAYWRIGHT = "Playwright",
  PHOTOSHOP = "Photoshop",
  REACT = "React",
  SASS = "Sass",
  SEO = "SEO",
  SLACK = "Slack",
  STYLED_COMPONENTS = "Styled Components",
  TYPESCRIPT = "TypeScript",
  UI = "UI",
  UX = "UX",
}

export type TechnologyData = {
  icon: string;
};